# src/swagger_server/models/new_user_response.py
# coding: utf-8

from __future__ import absolute_import

from src.swagger_server import Model
from src.swagger_server.models.user import User  # noqa: F401,E501
from src.swagger_server import util


class NewUserResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, user: User=None, message: str=None):  # noqa: E501
        """NewUserResponse - a model defined in Swagger

        :param user: The user of this NewUserResponse.  # noqa: E501
        :type user: User
        :param message: The message of this NewUserResponse.  # noqa: E501
        :type message: str
        """
        self.swagger_types = {
            'user': User,
            'message': str
        }

        self.attribute_map = {
            'user': 'user',
            'message': 'message'
        }

        self._user = user
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'NewUserResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The NewUserResponse of this NewUserResponse.  # noqa: E501
        :rtype: NewUserResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user(self) -> User:
        """Gets the user of this NewUserResponse.


        :return: The user of this NewUserResponse.
        :rtype: User
        """
        return self._user

    @user.setter
    def user(self, user: User):
        """Sets the user of this NewUserResponse.


        :param user: The user of this NewUserResponse.
        :type user: User
        """

        self._user = user

    @property
    def message(self) -> str:
        """Gets the message of this NewUserResponse.


        :return: The message of this NewUserResponse.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this NewUserResponse.


        :param message: The message of this NewUserResponse.
        :type message: str
        """

        self._message = message